MRMakeSlot.h
