+++ source/MRViewer/MRImGuiMenuListeners.cpp
@@ -1,6 +1,5 @@
 #include "MRImGuiMenuListeners.h"
-
-#include "MRViewer/MRViewer.h"
+#include "MRMakeSlot.h"
 
 namespace MR
 {
@@ -9,7 +8,7 @@ void NameTagClickListener::connect( Viewer* viewer, int group, boost::signals2::
 {
     if ( !viewer )
         return;
-    auto menu = viewer->getMenuPlugin();
+    auto menu = ImGuiMenu::instance();
     assert( menu );
     if ( menu )
         connection_ = menu->nameTagClickSignal.connect( group, MAKE_SLOT( &NameTagClickListener::onNameTagClicked_ ), pos );
@@ -19,7 +18,7 @@ void DrawSceneUiListener::connect( Viewer* viewer, int group, boost::signals2::c
 {
     if ( !viewer )
         return;
-    auto menu = viewer->getMenuPlugin();
+    auto menu = ImGuiMenu::instance();
     assert( menu );
     if ( menu )
         connection_ = menu->drawSceneUiSignal.connect( group, MAKE_SLOT( &DrawSceneUiListener::onDrawSceneUi_ ), pos );
